requirements.txt
Flask>=2.0.0
pyodbc>=4.0.0
pymongo>=4.0.0
bcrypt>=4.0.0
google-generativeai>=0.3.0

